Refactoring slice
1. slice.js에서 reducer를 export default로 변경하였습니다.
2. slice.js에서 직접 actions를 export하도록 변경하고 actions.js에서는 비동기 action을 제외한 action은 export하지 않도록 변경하였습니다.
3. action을 사용하는 component와 test에서 직접 slice에서 가져와 사용하도록 변경하였습니다.
