+++ src/actions.js
@@ -9,22 +9,15 @@ import {
 
 import { saveItem } from './services/storage';
 
-import { actions } from './slice';
-
-export const {
+import {
   setRegions,
   setCategories,
   setRestaurants,
   setRestaurant,
-  selectRegion,
-  selectCategory,
-  changeLoginField,
   setAccessToken,
-  logout,
-  changeReviewField,
   clearReviewFields,
   setReviews,
-} = actions;
+} from './slice';
 
 export function loadInitialData() {
   return async (dispatch) => {
