+++ src/reducer.test.js
@@ -1,8 +1,4 @@
-import reducer from './reducer';
-
-import { actions } from './slice';
-
-const {
+import reducer, {
   setRegions,
   setCategories,
   setRestaurants,
@@ -15,7 +11,7 @@ const {
   changeReviewField,
   clearReviewFields,
   setReviews,
-} = actions;
+} from './slice';
 
 describe('reducer', () => {
   context('when previous state is undefined', () => {
