+++ src/slice.js
@@ -127,7 +127,19 @@ const { actions, reducer } = createSlice({
   },
 });
 
-export {
-  actions,
-  reducer,
-};
+export const {
+  setRegions,
+  setCategories,
+  setRestaurants,
+  setRestaurant,
+  selectRegion,
+  selectCategory,
+  changeLoginField,
+  setAccessToken,
+  logout,
+  changeReviewField,
+  clearReviewFields,
+  setReviews,
+} = actions;
+
+export default reducer;
